docker: use multi-stage builds to run gradle
This PR adds the gradle build step to the Dockerfile. This way, contributors don't need to set up JDK to build the Docker image.

Because we use a separate build stage, the size of the image is the same as what's currently available on Docker Hub (479MB according to `docker image ls`)